requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy[asyncio]==2.0.29
asyncmy==0.2.9
alembic==1.13.1
redis==5.0.1
argon2-cffi==23.1.0
PyJWT[crypto]==2.8.0
fastapi-mail==1.4.1
pydantic[email]==2.6.4
python-dotenv==1.0.1
